refactor implementation of mixed
genes and mixed function are implemented as wrapper of a more general parse function that can handle file/folder and different combination of options
